Separating OLS and GLS algorithms into their own files, separate from the main ramp fitting module.
