+++ jwst/ramp_fitting/tests/test_ramp_fit.py
@@ -2,7 +2,7 @@ import pytest
 import numpy as np
 
 from jwst.ramp_fitting.ramp_fit import ramp_fit
-from jwst.ramp_fitting.ramp_fit import calc_num_seg
+from jwst.ramp_fitting.ols_fit import calc_num_seg
 from jwst.datamodels import dqflags
 from jwst.datamodels import RampModel
 from jwst.datamodels import GainModel, ReadnoiseModel
